refactor: 関数を整理
