+++ includes/Webserv.hpp
@@ -1,7 +1,6 @@
 #ifndef INCLUDES_WEBSERV_HPP
 #define INCLUDES_WEBSERV_HPP
 
-void server();
 void server_io_multiplexing();
 
 #endif /* INCLUDES_WEBSERV_HPP */
