+++ srcs/main.cpp
@@ -7,7 +7,6 @@ int main(int argc, char **argv) {
     std::cerr << "Usage: ./webserv <config file>" << std::endl;
     return (1);
   }
-  // server();
   server_io_multiplexing();
   return (0);
 }
