+++ srcs/message/response_message.cpp
@@ -1,6 +1,42 @@
 #include "message.hpp"
+#include "util.hpp"
 #include <iostream>
 #include <map>
+#include <sys/socket.h>
+#include <unistd.h>
+
+#define BUF_SIZE 1024
+
+std::string read_request(int accfd) {
+  char        buf[BUF_SIZE] = {};
+  std::string recv_str      = "";
+  ssize_t     read_size     = 0;
+
+  do {
+    read_size = recv(accfd, buf, sizeof(buf) - 1, 0);
+    if (read_size == -1) {
+      error_log_with_errno("read() failed.");
+      close(accfd);
+      // accfd = -1;
+      break;
+    }
+    if (read_size > 0) {
+      recv_str.append(buf);
+    }
+    if (is_match_suffix_string(recv_str, "\r\n\r\n")) {
+      break;
+    }
+  } while (read_size > 0);
+  std::cout << "read_request()" << std::endl;
+  std::cout << recv_str << std::endl;
+  return recv_str;
+}
+
+void send_response(int accfd, const std::string &message) {
+  send(accfd, message.c_str(), message.size(), 0);
+  std::cout << "send_response()" << std::endl;
+  std::cout << message << std::endl;
+}
 
 /*
  * ステータスラインの要素は必須だが, 存在しなかった時のバリデートは現状してない
